Adds web entrypoint
